% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee$Image$Extra_getCitation(x)}
\alias{ee$Image$Extra_getCitation(x)}
\alias{ee_image_getCitation}
\title{Citing EE objects in publications}
\usage{
`ee$Image$Extra_getCitation()`
}
\arguments{
\item{x}{An EE object to get the citation from.}
}
\value{
A character with citation information.
}
\description{
If it exists, retrieve the citation of an EE object.
}
\examples{
\dontrun{
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection("NASA/GPM_L3/IMERG_V06") \%>\%
  ee$ImageCollection$first() \%>\%
  ee$Image$Extra_getCitation()
}
}
\concept{ee_Image}
