% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R, R/ee_ImageColletion.R
\name{ee-getdoi}
\alias{ee-getdoi}
\alias{ee_image_getDOI}
\alias{ee_ImageCollection_getDOI}
\title{Get the Digital Object Identifier (DOI) of an EE object}
\usage{
`ee$Image$Extra_getDOI(x)`

`ee$ImageCollection$Extra_getDOI(x)`
}
\arguments{
\item{x}{An EE object.}
}
\value{
A character with DOI information.
}
\description{
If it exists, retrieve the DOI of an EE object.
}
\examples{
\dontrun{
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection("NASA/GPM_L3/IMERG_V06") \%>\%
  ee$ImageCollection$first() \%>\%
  ee$Image$Extra_getDOI()
}
}
