% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R, R/ee_ImageCollection.R
\name{ee-getoffset}
\alias{ee-getoffset}
\alias{ee_Image_getOffsetParams}
\alias{ee_ImageCollection_getOffsetParams}
\title{Retrieve offset parameter from EE Image and EE ImageCollections objects}
\usage{
`ee$Image$Extra_getOffsetParams(x)`

`ee$ImageCollection$Extra_getOffsetParams(x)`
}
\arguments{
\item{x}{An ee$Image or an ee$ImageCollection object.}
}
\value{
A list with the offset parameters for each band.
}
\description{
Earth Engine apply a simply lossless compression technique: IMG_Float_Values =
scale * IMG_Integer_Values + offset. ee$Image$getOffsetParams or
ee$ImageCollection$getOffsetParams retrieve the offset parameter
for each band of an ee$Image.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

# Retrieve offset parameters from the first image in NASA IMERG V06 collection
offset_params <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] \%>\%
  ee$Image$Extra_getOffsetParams()

# Display offset parameters for each band.
offset_params

# Get offset parameters from NASA IMERG V06 image collection.
offset_params_ <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06") \%>\%
  ee$ImageCollection$Extra_getOffsetParams()

# Display offset parameters for each band.
offset_params_
}
}
