% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee-pansharpen}
\alias{ee-pansharpen}
\alias{ee_Image_panSharpen}
\title{Apply panchromatic sharpening to an ee$Image or an ee$ImageCollection}
\usage{
`ee$Image$Extra_panSharpen(x, ...)`
}
\arguments{
\item{x}{An ee$Image object, the image to sharpen.}

\item{...}{Additional arguments including sharpening method, quality
assessments, and parameters for \code{ee.Image.reduceRegion()}.
See details for more information.}
}
\value{
The Image with all sharpenable bands sharpened to the
panchromatic resolution and quality assessments run and set
as properties.
}
\description{
Apply panchromatic sharpening to an ee$Image. Optionally,
run quality assessments between the original and
sharpened Image to measure spectral distortion and set results
as properties of the sharpened Image.
}
\details{
The \code{...} argument can include the following:
\itemize{
\item{method}{Character. The sharpening algorithm to apply. Options include “SFIM”,
“HPFA”, “PCS”, and “SM”. Default is “SFIM”.}
\item{qa}{Character. One or more quality assessment names to apply after sharpening,
such as “MSE”, “RASE”, “UIQI”, etc.}
\item{geometry, maxPixels, bestEffort, etc.}{Arguments passed to \code{ee.Image.reduceRegion()}
during PCS sharpening and quality assessments.}
}
For the PCS method, additional parameters for \code{ee.Image.reduceRegion()} can be specified,
such as \code{geometry}, \code{maxPixels}, \code{bestEffort}, etc.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

img <- ee$Image("LANDSAT/LC09/C02/T1_TOA/LC09_047027_20230815")
img_sharp <- ee$Image$Extra_panSharpen(img, method="HPFA", qa=c("MSE", "RMSE"), maxPixels=1e13)

Map$centerObject(img)
Map$addLayer(img_sharp, list(bands=c("B4", "B3", "B2"))) |
Map$addLayer(img, list(bands=c("B4", "B3", "B2")))
}
}
