% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R, R/ee_ImageCollection.R
\name{ee-preprocess}
\alias{ee-preprocess}
\alias{ee_Image_preprocess}
\alias{ee_ImageCollection_preprocess}
\title{Automated EE Image or EE ImageCollection preprocessing}
\usage{
`ee$Image$Extra_preprocess(x, ...)`

`ee$ImageCollection$Extra_preprocess(x, ...)`
}
\arguments{
\item{x}{An ee$Image or an ee$ImageCollection object.}

\item{...}{Arguments to pass to ee$Image$cloudmask.}
}
\value{
An ee$Image or ee$ImageCollection object
}
\description{
Preprocessing of ee$Image or ee$ImageCollection objects. This
function performs the following tasks:
\itemize{
\item \strong{Cloud Masking}: Remove cloud and cloud shadow pixels.
See ee$Image$cloudmask.
\item \strong{Decompress}: Convert integer pixels to float point numbers.
}
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

# Load a Sentinel-2 image and apply automated preprocessing.
img <- ee$Image("COPERNICUS/S2_SR/20170328T083601_20170328T084228_T35SQA") \%>\%
  ee$Image$Extra_preprocess()

# Load and preprocess Sentinel-2 SR image collection.
ic <- ee$ImageCollection$Dataset$COPERNICUS_S2_SR \%>\%
  ee$ImageCollection$Extra_preprocess()
}
}
