% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R, R/ee_ImageCollection.R
\name{ee-scaleandoffset}
\alias{ee-scaleandoffset}
\alias{ee_Image_scaleAndOffset}
\alias{ee_ImageCollection_scaleAndOffset}
\title{Automatic decompression of EE Image or EE ImageCollection objects}
\usage{
`ee$Image$Extra_scaleAndOffset(x)`

`ee$ImageCollection$Extra_scaleAndOffset(x)`
}
\arguments{
\item{x}{An ee$Image or an ee$ImageCollection object.}
}
\value{
An ee$Image or an ee$ImageCollection with float pixel values.
}
\description{
Earth Engine apply a simply lossless compression technique: IMG_Float_Values =
scale * IMG_Integer_Values + offset. ee$Image$scaleAndOffset or
ee$ImageCollection$scaleAndOffset backs the integer pixel values to
float point number.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

# Adjust first image in NASA IMERG V06 for scale and offset.
adjusted_image <- ee$ImageCollection("NASA/GPM_L3/IMERG_V06")[[1]] \%>\%
  ee$Image$Extra_scaleAndOffset()

# Adjust Sentinel-2 SR images for scale and offset.
adjusted_images <- ee$ImageCollection("COPERNICUS/S2_SR") \%>\%
  ee$ImageCollection$Extra_scaleAndOffset()
}
}
\concept{calibration}
