% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R, R/ee_ImageColletion.R
\name{ee-tasseledcap}
\alias{ee-tasseledcap}
\alias{ee_Image_tasseledCap}
\alias{ee_ImageCollection_tasseledCap}
\title{Calculates tasseled cap brightness, wetness, and greenness
components for an EE Image or EE ImageCollection}
\usage{
`ee$Image$Extra_tasseledCap()`

`ee$ImageCollection$Extra_tasseledCap()`
}
\arguments{
\item{x}{ee$Image or ee$ImageCollection to calculate tasseled
cap components for. Must belong to a supported platform.}
}
\value{
ee$Image  or ee$ImageCollection with the tasseled cap
components as new bands.
}
\description{
Tasseled cap transformations are applied using coefficients
published for these supported platforms:
\itemize{
\item{1.} Sentinel-2 MSI Level 1C  (1)
\item{2.} Landsat 9 OLI-2 SR  (2)
\item{3.} Landsat 9 OLI-2 TOA  (2)
\item{4.} Landsat 8 OLI SR  (2)
\item{5.} Landsat 8 OLI TOA  (2)
\item{6.} Landsat 7 ETM+ TOA  (3)
\item{7.} Landsat 5 TM Raw DN  (4)
\item{8.} Landsat 4 TM Raw DN  (5)
\item{9.} Landsat 4 TM Surface Reflectance  (6)
\item{10.} MODIS NBAR  (7)
}
}
\details{
\itemize{
\item{1.} Shi, T., & Xu, H. (2019). Derivation of Tasseled Cap Transformation
Coefficients for Sentinel-2 MSI At-Sensor Reflectance Data. IEEE Journal
of Selected Topics in Applied Earth Observations and Remote Sensing, 1–11.
doi:10.1109/jstars.2019.2938388
\item{2.} Zhai, Y., Roy, D.P., Martins, V.S., Zhang, H.K., Yan, L., Li, Z. 2022.
Conterminous United States Landsat-8 top of atmosphere and surface reflectance
tasseled cap transformation coefficeints. Remote Sensing of Environment,
274(2022). doi:10.1016/j.rse.2022.112992
\item{3.} Huang, C., Wylie, B., Yang, L., Homer, C. and Zylstra, G., 2002.
Derivation of a tasselled cap transformation based on Landsat 7 at-satellite
reflectance. International journal of remote sensing, 23(8), pp.1741-1748.
\item{4.} Crist, E.P., Laurin, R. and Cicone, R.C., 1986, September. Vegetation
and soils information contained in transformed Thematic Mapper data. In
Proceedings of IGARSS’86 symposium (pp. 1465-1470). Paris: European Space
Agency Publications Division.
\item{5.} Crist, E.P. and Cicone, R.C., 1984. A physically-based transformation
of Thematic Mapper data---The TM Tasseled Cap. IEEE Transactions on Geoscience
and Remote sensing, (3), pp.256-263.
\item{6.} Crist, E.P., 1985. A TM tasseled cap equivalent transformation for
reflectance factor data. Remote sensing of Environment, 17(3), pp.301-306.
\item{7.} Lobser, S.E. and Cohen, W.B., 2007. MODIS tasselled cap: land cover
characteristics expressed through transformed MODIS data. International
Journal of Remote Sensing, 28(22), pp.5079-5101.
}
}
\examples{
\dontrun{
library(rgeeExtra)

ee_Initialize()

img <- ee$Image("LANDSAT/LT05/C01/T1/LT05_044034_20081011")
img <- ee$Image$Extra_tasseledCap(img)
names(img)
}
}
