% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee_Image_getCitation}
\alias{ee_Image_getCitation}
\title{Citing EE Image objects in publications}
\usage{
ee_Image_getCitation(x)
}
\arguments{
\item{x}{Image to get the citation from.}
}
\value{
A character with citation information.
}
\description{
If it exists, retrieve the citation of an ee$Image object.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection$Dataset$NASA_GPM_L3_IMERG_V06$first() \%>\%
  ee_Image_getCitation()
}
}
\seealso{
Other citation: 
\code{\link{ee_ImageCollection_getCitation}()},
\code{\link{ee_ImageCollection_getDOI}()}
}
\concept{citation}
