% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee_Image_getOffsetParams}
\alias{ee_Image_getOffsetParams}
\title{Retrieve offset parameter from EE Image objects}
\usage{
ee_Image_getOffsetParams(x)
}
\arguments{
\item{x}{ee$Image}
}
\value{
A list with the offset parameters for each band.
}
\description{
Earth Engine apply a simply lossless compression technique: IMG_Float_Values =
scale * IMG_Integer_Values + offset. ee_Image_getOffsetParams
retrieve the offset parameter for each band of an ee$Image.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection$Dataset$NASA_GPM_L3_IMERG_V06$first() \%>\%
  ee_Image_getOffsetParams()
}
}
\seealso{
Other calibration: 
\code{\link{ee_ImageCollection_getOffsetParams}()},
\code{\link{ee_ImageCollection_getScaleParams}()},
\code{\link{ee_ImageCollection_scaleAndOffset}()},
\code{\link{ee_Image_getScaleParams}()}
}
\concept{calibration}
