% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee_Image_getSTAC}
\alias{ee_Image_getSTAC}
\title{Retrieve EE Image STAC metadata}
\usage{
ee_Image_getSTAC(x)
}
\arguments{
\item{x}{ee$Image.}
}
\value{
Return STAC metadata for each band.
}
\description{
Get the STAC of an ee$Image object.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection$Dataset$NASA_GPM_L3_IMERG_V06$first() \%>\%
  ee_Image_getSTAC()
}
}
\seealso{
Other STAC: 
\code{\link{ee_ImageCollection_getSTAC}()}
}
\concept{STAC}
