% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee_Image_matchHistogram}
\alias{ee_Image_matchHistogram}
\title{Adjust the image's histogram to match a target image}
\usage{
`ee$Image$Extra_matchHistogram(x, ...)`
}
\arguments{
\item{x}{ee$Image to adjust.}

\item{...}{Additional arguments for histogram matching.
See details for more information.}
}
\value{
The adjusted image containing the matched source bands.
}
\description{
Matches the histogram of one image (source) to that of another image (target).
}
\details{
The \code{...} argument can include the following:
\itemize{
\item{target}{ee$Image. The target image to match.}
\item{bands}{Dictionary. Band names to match, with source bands as keys and target bands as values.}
\item{geometry}{ee$Geometry. The region to match histograms in that overlaps both images. Default is NULL.}
\item{maxBuckets}{Integer. The maximum number of buckets to use when building histograms. Default 256.}
}
These parameters allow for detailed customization of the histogram matching process.
}
\examples{
\dontrun{
library(rgeeExtra)

ee_Initialize()

source <- ee$Image("LANDSAT/LC08/C01/T1_TOA/LC08_047027_20160819")
target <-ee$Image("LANDSAT/LE07/C01/T1_TOA/LE07_046027_20150701")
bands <- list("B4"="B3", "B3"="B2", "B2"="B1")

matched <- ee$Image$Extra_matchHistogram(source, target, bands)

names(matched)
}
}
