% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee_Image_preprocess}
\alias{ee_Image_preprocess}
\title{Automated EE Image preprocessing}
\usage{
ee_Image_preprocess(x, ...)
}
\arguments{
\item{x}{ee$Image.}

\item{...}{Arguments to pass to ee_Image_preprocess.}
}
\value{
An ee$Image object
}
\description{
Preprocessing of ee$Image objects. This function performs the
following tasks:
\itemize{
\item \strong{Cloud Masking}: Remove cloud and cloud shadow pixels.
See ee_model_cloudmask.
\item \strong{Decompress}: Convert integer pixels to float point numbers.
}
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection$Dataset$COPERNICUS_S2_SR \%>\%
  ee$ImageCollection$first() \%>\%
  ee_Image_preprocess()
}
}
