% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee_Image_spectralIndex}
\alias{ee_Image_spectralIndex}
\title{Spectral Indices Computation}
\usage{
`ee$Image$Extra_spectralIndex(x, ...)`
}
\arguments{
\item{x}{An ee$Image or an ee$ImageCollection to compute indices on.}

\item{...}{Additional arguments passed to the underlying spectral index
computation function. See details for a complete list of possible arguments.}
}
\value{
An ee$Image or an ee$ImageCollection with the computed spectral
index, or indices, as new bands.
}
\description{
Computes one or more spectral indices for an ee$Image or an ee$ImageCollection object.
}
\details{
The \code{...} argument can include any of the following:
\itemize{
\item{index}{Character. Index or list of indices to compute. Options include 'vegetation', 'burn',
'water', 'snow', 'drought', 'urban', 'kernel', and 'all'. Default 'NDVI'.}
\item{G}{Numeric. Gain factor for 'EVI'. Default 2.5.}
\item{C1, C2}{Numerics. Coefficients for aerosol resistance in 'EVI'. Defaults are 6.0 and 7.5.}
\item{L}{Numeric. Canopy background adjustment for 'EVI' or 'SAVI'. Default 1.0.}
\item{cexp}{Numeric. Coefficient for 'OCVI'. Default 1.16.}
\item{nexp}{Numeric. Exponent for 'GDVI'. Default 2.0.}
\item{alpha}{Numeric. Weighting coefficient for 'WDRVI'. Default 0.1.}
\item{slope, intercept}{Numerics. Soil line slope and intercept. Defaults are 1.0 and 0.0.}
\item{gamma}{Numeric. Weighting coefficient for 'ARVI'. Default 1.0.}
\item{kernel}{Character. Kernel type for kernel indices. Options are 'linear', 'RBF', and 'poly'. Default 'RBF'.}
\item{sigma}{Character or Numeric. Length-scale parameter for RBF kernel. Default '0.5 * (a + b)'.}
\item{p}{Numeric. Kernel degree for polynomial kernel. Default 2.0.}
\item{c}{Numeric. Free parameter for polynomial kernel. Default 1.0.}
\item{online}{Logical. Whether to retrieve the most recent list of indices online. Default FALSE.}
\item{drop}{Logical. If TRUE, drop the image bands after calculation. Default TRUE.}
}
For a complete list of indices and their parameters, please refer to the
\href{https://awesome-ee-spectral-indices.readthedocs.io/en/latest/list.html}{spectral indices documentation}.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

s2_indices <- ee$ImageCollection("COPERNICUS/S2_SR") \%>\%
  ee$ImageCollection$first() \%>\%
  ee$Image$Extra_preprocess() \%>\%
  ee$Image$Extra_spectralIndex(c("NDVI", "SAVI"))

names(s2_indices)
# "NDVI" "SAVI"
}
}
