% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_ImageColletion.R
\name{ee_ImageCollection_closest}
\alias{ee_ImageCollection_closest}
\title{Get the temporal nearest image}
\usage{
`ee$ImageCollection$Extra_closest(x, ...)`
}
\arguments{
\item{x}{ee$ImageCollection from which to get the closest image to the specified date.}

\item{...}{Additional arguments for finding the nearest image.
See details for more information.}
}
\value{
An ee$Image closest to the specified date.
}
\description{
Gets the closest ee$Image to a specified date from an ee$ImageCollection.
}
\details{
The \code{...} argument can include the following:
\itemize{
\item{date}{ee$Date or R date object. Date of interest for searching the closest image.}
\item{tolerance}{Numeric. Filters the collection within a range of (date - tolerance, date + tolerance). Default 1.}
\item{unit}{Character. Units for tolerance. Options include "year", "month", "week", "day", "hour", "minute", "second". Default "month".}
}
These parameters allow for precise control over how the closest image is determined.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

roi <- ee$Geometry$Point(c(-79, -12))
ee$ImageCollection$Dataset$MODIS_006_MCD12Q1 \%>\%
  ee$ImageCollection$Extra_closest("2020-10-15",  2, "year") \%>\%
  ee$ImageCollection$first() \%>\%
  Map$addLayer()
}
}
