% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_ImageColletion.R
\name{ee_ImageCollection_getDOI}
\alias{ee_ImageCollection_getDOI}
\title{Get the Digital Object Identifier (DOI) of an EE ImageCollection object}
\usage{
ee_ImageCollection_getDOI(x)
}
\arguments{
\item{x}{ee$ImageCollection.}
}
\value{
A character with DOI information.
}
\description{
If it exists, retrieve the DOI of an ee$ImageCollection object.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection$Dataset$NASA_GPM_L3_IMERG_V06 \%>\%
  ee_ImageCollection_getDOI()
}
}
\concept{citation}
