% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_ImageColletion.R
\name{ee_ImageCollection_getScaleParams}
\alias{ee_ImageCollection_getScaleParams}
\title{Retrieve scale parameter from ee$ImageCollection objects.}
\usage{
ee_ImageCollection_getScaleParams(x)
}
\arguments{
\item{x}{ee$ImageCollection.}
}
\value{
A list with the scale parameters for each band.
}
\description{
Earth Engine apply a simply lossy compression technique: IMG_Float_Values =
scale * IMG_Integer_Values + offset. ee_ImageCollection_getScaleParams
retrieve the scale parameter for each band of an ee$ImageCollection.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection$Dataset$NASA_GPM_L3_IMERG_V06 \%>\%
  ee_ImageCollection_getScaleParams()
}
}
