% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_ImageColletion.R
\name{ee_ImageCollection_getSTAC}
\alias{ee_ImageCollection_getSTAC}
\title{Retrieve EE ImageCollection STAC metadata}
\usage{
ee_ImageCollection_getSTAC(x)
}
\arguments{
\item{x}{ee$ImageCollection.}
}
\value{
Return STAC metadata for each band.
}
\description{
Get the STAC of an ee$ImageCollection object.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection$Dataset$NASA_GPM_L3_IMERG_V06 \%>\%
  ee_ImageCollection_getSTAC()
}
}
\seealso{
Other STAC: 
\code{\link{ee_Image_getSTAC}()}
}
\concept{STAC}
