% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_ImageColletion.R
\name{ee_ImageCollection_scaleAndOffset}
\alias{ee_ImageCollection_scaleAndOffset}
\title{Retrieve real values from ee$ImageCollection objects}
\usage{
ee_ImageCollection_scaleAndOffset(x)
}
\arguments{
\item{x}{: ee$ImageCollection.}
}
\value{
An ee$ImageCollection with float pixel values.
}
\description{
Earth Engine apply a simply lossy compression technique: IMG_Float_Values =
scale * IMG_Integer_Values + offset. ee_ImageCollection_scaleAndOffset backs
the integer ee$Image pixel values to float point number.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection$Dataset$COPERNICUS_S2_SR \%>\%
  ee$ImageCollection$first() \%>\%
  ee_ImageCollection_scaleAndOffset()
}
}
