% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-imagecollection.R
\name{ee_length_ic}
\alias{ee_length_ic}
\alias{length.ee.imagecollection.ImageCollection}
\title{Length of an Earth Engine ImageCollection Object}
\usage{
\method{length}{ee.imagecollection.ImageCollection}(x)
}
\arguments{
\item{x}{an EE ImageCollection Object.}
}
\description{
Set the length of an Earth Engine Image.
}
\examples{
\dontrun{
library(rgeeExtra)
library(rgee)

ee_Initialize()     # Initialize the Google Earth Engine API connection
extra_Initialize()  # Initialize the extended functionalities of rgeeExtra

ic <- ee$ImageCollection("COPERNICUS/S2_SR") \%>\%
  ee$ImageCollection$filterDate("1999-01-01", "1999-01-02")
length(ic)
}
}
