% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-imagecollection.R
\name{ee_name_ic}
\alias{ee_name_ic}
\alias{names.ee.imagecollection.ImageCollection}
\title{Names of Earth Engine ImagesCollection properties}
\usage{
\method{names}{ee.imagecollection.ImageCollection}(x)
}
\arguments{
\item{x}{an EE ImageCollection object.}
}
\description{
Get the names of the properties of an Earth Engine ImageCollection object.
}
