% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-fc.R, R/ops-subsetting-image.R,
%   R/ops-subsetting-imagecollection.R
\name{ee_subsetting}
\alias{ee_subsetting}
\alias{[[.ee.featurecollection.FeatureCollection}
\alias{names.ee.feature.Feature}
\alias{[[.ee.image.Image}
\alias{[<-.ee.image.Image}
\alias{[[<-.ee.image.Image}
\alias{[[.ee.imagecollection.ImageCollection}
\alias{[[<-.ee.imagecollection.ImageCollection}
\alias{names.ee.imagecollection.ImageCollection}
\title{Extract parts of and EE FeatureCollection}
\usage{
\method{[[}{ee.featurecollection.FeatureCollection}(x, index)

\method{names}{ee.feature.Feature}(x)

\method{[[}{ee.image.Image}(x, index)

\method{[}{ee.image.Image}(x, index) <- value

\method{[[}{ee.image.Image}(x, index) <- value

\method{[[}{ee.imagecollection.ImageCollection}(x, index)

\method{[[}{ee.imagecollection.ImageCollection}(x, index) <- value

\method{names}{ee.imagecollection.ImageCollection}(x)
}
\arguments{
\item{x}{an EE ImageCollection object.}

\item{index}{Integer. Index specifying elements to extract or replace.}

\item{value}{ee$ImageCollection or ee$Image to replace in.}
}
\value{
An ee$FeatureCollection

A vector representing the property names of the ee$FeatureCollection object

An EE.ImageCollection

An EE.ImageCollection

A vector with the property names of ee$ImageCollection object
}
\description{
Get the properties names of FeatureCollection object

Get the names of the properties of an Earth Engine ImageCollection object.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)
library(sf)

ee_Initialize(gcs = TRUE, drive = TRUE)
extra_Initialize()

# Define a Image or ImageCollection: Terraclimate
fc_tiger <- ee$FeatureCollection('TIGER/2016/Roads')
fc_tiger_subset <- fc_tiger[[1:100]]
Map$centerObject(fc_tiger_subset)
Map$addLayer(fc_tiger_subset)
}
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

fc <- ee$FeatureCollection('WRI/GPPD/power_plants')
fc$propertyNames()$getInfo()
}
\dontrun{
library(rgee)
library(rgeeExtra)
library(sf)

ee_Initialize(gcs = TRUE, drive = TRUE)

# Define a Image or ImageCollection: Terraclimate
terraclimate <- ee$ImageCollection("IDAHO_EPSCOR/TERRACLIMATE") \%>\%
  ee$ImageCollection$filterDate("2001-01-01", "2002-01-01")

# Define temperature Vis parameters
maximumTemperatureVis <- list(
  min = -300.0,
  max = 300.0,
  palette = c(
    '1a3678', '2955bc', '5699ff', '8dbae9', 'acd1ff', 'caebff', 'e5f9ff',
    'fdffb4', 'ffe6a2', 'ffc969', 'ffa12d', 'ff7c1f', 'ca531a', 'ff0000',
    'ab0000'
  )
)

Map$setCenter(71.72, 52.48, 2)
tnames <- names(terraclimate[[2]])
m1 <- Map$addLayer(terraclimate[[2]][["tmmx"]], maximumTemperatureVis)

terraclimate[[2]] <- terraclimate[[2]]*1.4
names(terraclimate[[2]]) <- tnames
m2 <- Map$addLayer(terraclimate[[2]][["tmmx"]], maximumTemperatureVis)
m1 | m2
}
\dontrun{
library(rgee)
library(rgeeExtra)
library(sf)

ee_Initialize(gcs = TRUE, drive = TRUE)
extra_Initialize()
# Define a Image or ImageCollection: Terraclimate
}
\dontrun{
library(rgeeExtra)
library(rgee)

ee_Initialize()     # Initialize the Google Earth Engine API connection
extra_Initialize()  # Initialize the extended functionalities of rgeeExtra

ic <- ee$ImageCollection("COPERNICUS/S2_SR") \%>\%
  ee$ImageCollection$filterDate("1999-01-01", "1999-01-02")
names(ic)
}
}
