% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeeExtra-package.R
\docType{data}
\name{EEextra_PYTHON_PACKAGE}
\alias{EEextra_PYTHON_PACKAGE}
\title{Extra Earth Engine module}
\format{
Earth Engine module
}
\usage{
EEextra_PYTHON_PACKAGE
}
\description{
Interface to extent the Earth Engine package. Provides access to the
top level classes and functions.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)
ee_Initialize()
}
}
\keyword{datasets}
