% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeeExtra-package.R
\name{extra_Initialize}
\alias{extra_Initialize}
\title{Load extra functionality for rgee}
\usage{
extra_Initialize()
}
\description{
Load extra functionality for rgee
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()
}
}
