% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-image.R
\name{length<-.ee.image.Image}
\alias{length<-.ee.image.Image}
\title{Length of an Earth Engine Image Object}
\usage{
\method{length}{ee.image.Image}(x) <- value
}
\arguments{
\item{x}{an EE Image Object.}

\item{value}{A non-negative integer which increase or filter the initial lenght of the
ee$Image object.}
}
\value{
A filter or increase the number of bands in a ee$Image.
}
\description{
Get the length of an Earth Engine Image.
}
\examples{
\dontrun{
library(rgeeExtra)
library(rgee)

ee_Initialize()     # Initialize the Google Earth Engine API connection
extra_Initialize()  # Initialize the extended functionalities of rgeeExtra

ic <- ee$ImageCollection("COPERNICUS/S2_SR")$first()
length(ic) <- 10
}
}
