% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-fc.R
\name{length.ee.featurecollection.FeatureCollection}
\alias{length.ee.featurecollection.FeatureCollection}
\title{Length of an Earth Engine FeatureCollection Object}
\usage{
\method{length}{ee.featurecollection.FeatureCollection}(x)
}
\arguments{
\item{x}{an EE FeatureCollection Object.}
}
\value{
A numeric representing the length of a ee$FeatureCollection object.
}
\description{
Get the length of an Earth Engine FeatureCollection.
}
\examples{
\dontrun{

library(rgeeExtra)
library(rgee)

ee_Initialize()
extra_Initialize()

region <- ee$Geometry$Rectangle(-119.224, 34.669, -99.536, 50.064)
randomPoints <- ee$FeatureCollection$randomPoints(region)
length(randomPoints)
}
}
