% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-image.R
\name{length.ee.image.Image}
\alias{length.ee.image.Image}
\title{Length of an Earth Engine Image Object}
\usage{
\method{length}{ee.image.Image}(x)
}
\arguments{
\item{x}{an EE Image Object.}
}
\value{
A numeric value that indicate the number of bands in a ee$Image.
}
\description{
Get or set the length of an Earth Engine Image.
}
\details{
If a vector is shortened, extra values are discarded and when a vector
is lengthened, it is padded out to its new length with ee$Image(0), with
band name of zzz_rgee_NA_\%02d.
}
\examples{
\dontrun{
library(rgeeExtra)
library(rgee)

ee_Initialize()     # Initialize the Google Earth Engine API connection
extra_Initialize()  # Initialize the extended functionalities of rgeeExtra

ic <- ee$Image("COPERNICUS/S2_SR/20190310T105851_20190310T110327_T30TVK")
length(ic)
}
}
