% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_module.R
\name{module}
\alias{module}
\title{Load Javascript modules}
\usage{
module(x, upgrade = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{Character. Javascript module path.}

\item{upgrade}{Upgrade the js module}

\item{quiet}{Logical. Suppress info messages.}
}
\description{
Load Javascript modules
}
