% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_module.R
\name{module_uninstall}
\alias{module_uninstall}
\title{Delete a Js module from your local repository}
\usage{
module_uninstall(x)
}
\arguments{
\item{x}{Character. Javascript module path.}
}
\description{
Delete a Js module from your local repository
}
