% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-image.R
\name{names(x) <- value}
\alias{names(x) <- value}
\alias{names<-.ee.image.Image}
\title{Rename of Earth Engine Images layers (bands)}
\usage{
`names(x) <- value`
}
\arguments{
\item{x}{an EE Image object.}

\item{value}{a character vector with the same length as x.}
}
\value{
An ee$Image with its bands renamed.
}
\description{
Set the names of the layers of an Earth Engine Image object.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

img_demo <- ee$Image("COPERNICUS/S2_SR/20190310T105851_20190310T110327_T30TVK")[[1:3]]
names(img_demo) <- c("B01", "B02", "B03")
}
}
