% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-image.R
\name{names.ee.image.Image}
\alias{names.ee.image.Image}
\title{Names of Earth Engine Images layers (bands)}
\usage{
`names(x)`
}
\arguments{
\item{x}{an EE Image object.}
}
\value{
A vector with the name of the bands.
}
\description{
Get the names of the layers of an Earth Engine Image object.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

img_demo <- ee$Image("COPERNICUS/S2_SR/20190310T105851_20190310T110327_T30TVK")[[1:3]]
names(img_demo)
}
}
