% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-imagecollection.R
\name{names.ee.imagecollection.ImageCollection}
\alias{names.ee.imagecollection.ImageCollection}
\title{Names of Earth Engine ImagesCollection properties}
\usage{
\method{names}{ee.imagecollection.ImageCollection}(x)
}
\arguments{
\item{x}{an EE ImageCollection object.}
}
\value{
A vector with the property names of ee$ImageCollection object
}
\description{
Get the names of the properties of an Earth Engine ImageCollection object.
}
\examples{
\dontrun{
library(rgeeExtra)
library(rgee)

ee_Initialize()     # Initialize the Google Earth Engine API connection
extra_Initialize()  # Initialize the extended functionalities of rgeeExtra

ic <- ee$ImageCollection("COPERNICUS/S2_SR") \%>\%
  ee$ImageCollection$filterDate("1999-01-01", "1999-01-02")
names(ic)
}
}
