% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeeExtra-package.R
\docType{package}
\name{rgeeExtra-package}
\alias{rgeeExtra}
\alias{rgeeExtra-package}
\title{rgeeExtra: Extra functionality for rgee}
\description{
Google Earth Engine (Gorelick et al., 2017) is a cloud computing platform
designed for planetary-scale environmental data analysis that only can be
accessed via the Earth Engine code editor, third-party web apps, and the
JavaScript and Python client libraries. \code{rgee} is a non-official
client library for R that uses \code{reticulate} to wrap the Earth Engine
Python API and provide R users with a familiar interface, rapid development
features, and flexibility to analyze data using open-source, R third-party
packages.
}
\details{
The package implements and supports:

\itemize{
\item Math operators
}
}
\author{
\strong{Maintainer}: Cesar Aybar \email{csaybar@gmail.com} (\href{https://orcid.org/0000-0003-2745-9535}{ORCID})

Other contributors:
\itemize{
  \item David  Montero \email{dml.mont@gmail.com} (\href{https://orcid.org/0000-0002-9010-3286}{ORCID}) [contributor]
}

}
\keyword{package}
