% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting-imagecollection.R
\name{[[<-.ee.imagecollection.ImageCollection}
\alias{[[<-.ee.imagecollection.ImageCollection}
\title{Extract or replace parts of and ee$ImageCollection}
\usage{
\method{[[}{ee.imagecollection.ImageCollection}(x, index) <- value
}
\arguments{
\item{x}{ee$ImageCollection or ee$Image.}

\item{index}{Integer. Index specifying elements to extract or replace.}

\item{value}{ee$ImageCollection or ee$Image to replace in.}
}
\value{
An EE.ImageCollection
}
\description{
Extract or replace parts of and ee$ImageCollection
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)
library(sf)

ee_Initialize(gcs = TRUE, drive = TRUE)
extra_Initialize()
# Define a Image or ImageCollection: Terraclimate
}
}
