% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-math.R
\name{Summary-methods}
\alias{Summary-methods}
\alias{Summary.ee.image.Image}
\alias{mean.ee.image.Image}
\title{Summary Methods}
\usage{
\method{Summary}{ee.image.Image}(..., na.rm = TRUE)

\method{mean}{ee.image.Image}(..., na.rm = TRUE)
}
\arguments{
\item{...}{ee$Image.}

\item{na.rm}{Ignore.}
}
\value{
An ee$Image object
}
\description{
Summary Methods
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

roi <- ee$Geometry$Point(c(-79, -12))

img_demo <- ee$Image("COPERNICUS/S2_SR/20190310T105851_20190310T110327_T30TVK")
max(img_demo)
}
}
