% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R, R/ee_ImageColletion.R
\name{ee-citation}
\alias{ee-citation}
\alias{ee_image_getCitation}
\alias{ee_ImageCollection_getCitation}
\title{Citing EE objects in publications}
\usage{
ee_image_getCitation(x)

ee_ImageCollection_getCitation(x)
}
\arguments{
\item{x}{An EE object to get the citation from.}
}
\value{
A character with citation information.
}
\description{
If it exists, retrieve the citation of an EE object.
}
\examples{
\dontrun{
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection("NASA/GPM_L3/IMERG_V06") \%>\%
  ee$ImageCollection$first() \%>\%
  ee$Image$getCitation()
}
}
