% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R, R/ee_ImageColletion.R
\name{ee-getscaleparams}
\alias{ee-getscaleparams}
\alias{ee_Image_getScaleParams}
\alias{ee_ImageCollection_getScaleParams}
\title{Retrieve scale parameter from EE Image and EE ImageCollection objects.}
\usage{
ee_Image_getScaleParams(x)

ee_ImageCollection_getScaleParams(x)
}
\arguments{
\item{x}{An ee$Image.or an ee$ImageCollection object.}
}
\value{
A list with the scale parameters for each band.
}
\description{
Earth Engine apply a simply lossless compression technique: IMG_Float_Values =
scale * IMG_Integer_Values + offset. ee$Image$getScaleParams and
ee$ImageCollection$getScaleParams retrieve the scale parameter
for each band of an ee$Image.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection("NASA/GPM_L3/IMERG_V06") \%>\%
  ee$ImageCollection$first() \%>\%
  ee$Image$getScaleParams()
}
}
