% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R, R/ee_ImageColletion.R
\name{ee-getstac}
\alias{ee-getstac}
\alias{ee_Image_getSTAC}
\alias{ee_ImageCollection_getSTAC}
\title{Retrieve EE Image or EE ImageCollection STAC metadata}
\usage{
`ee$Image$Extra_getSTAC()`

`ee$ImageCollection$Extra_getSTAC()`
}
\arguments{
\item{x}{An ee$Image or an ee$ImageCollection object.}
}
\value{
Return STAC metadata for each band.
}
\description{
Get the STAC of an ee$Image or ee$ImageCollection object.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection("NASA/GPM_L3/IMERG_V06") \%>\%
  ee$ImageCollection$first() \%>\%
  ee$Image$Extra_getSTAC() \%>\%
  ee$Image$getInfo()
}
}
