% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R, R/ee_ImageColletion.R
\name{ee-pansharpen}
\alias{ee-pansharpen}
\alias{ee_Image_panSharpen}
\alias{ee_ImageCollection_panSharpen}
\title{Apply panchromatic sharpening to an ee$Image or an ee$ImageCollection}
\usage{
ee_Image_panSharpen(x, method = "SFIM", qa = "MSE", ...)

ee_ImageCollection_panSharpen(x, method = "SFIM", qa = "MSE", ...)
}
\arguments{
\item{x}{An ee$Image or an ee$ImageCollection object. The image to sharpen.}

\item{method}{Character. The sharpening algorithm to apply. Current
options are “SFIM” (Smoothing Filter-based Intensity Modulation),
“HPFA” (High Pass Filter Addition), “PCS” (Principal Component
Substitution), and “SM” (simple mean). Different sharpening
methods will produce different quality sharpening results in
different scenarios.}

\item{qa}{Character. One or more optional quality assessment names to
apply after sharpening, e.g. “MSE”, “RASE”, “UIQI”, etc.}

\item{...}{Keyword arguments passed to ee.Image.reduceRegion() such
as “geometry”, “maxPixels”, “bestEffort”, etc. These arguments are
only used for PCS sharpening and quality assessments.}
}
\value{
The Image with all sharpenable bands sharpened to the
panchromatic resolution and quality assessments run and set
as properties.
}
\description{
Apply panchromatic sharpening to an ee$Image or an ee$ImageCollection. Optionally,
run quality assessments between the original and
sharpened Image to measure spectral distortion and set results
as properties of the sharpened Image.
}
\examples{
\dontrun{
library(rgeeExtra)

ee_Initialize()

img <- ee$Image("LANDSAT/LC08/C01/T1_TOA/LC08_047027_20160819")
img_sharp <- ee$Image$panSharpen(img, method="HPFA", qa=c("MSE", "RMSE"), maxPixels=1e13)

Map$centerObject(img)
Map$addLayer(img_sharp, list(bands=c("B4", "B3", "B2"))) |
Map$addLayer(img, list(bands=c("B4", "B3", "B2")))
}
}
