% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee_Image_maskClouds}
\alias{ee_Image_maskClouds}
\title{Masks clouds and shadows}
\usage{
`ee$Image$Extra_maskClouds(x, ...)`
}
\arguments{
\item{x}{An ee$Image to be processed for cloud and shadow masking.}

\item{...}{Additional arguments for cloud and shadow masking.
See details for more information.}
}
\value{
ee$Image with a Cloud-shadow masked image.
}
\description{
Masks clouds and shadows in an ee.Image. Valid just for
Surface Reflectance products. This function may mask water
as well as clouds for the Sentinel-3 Radiance product.
}
\details{
The \code{...} argument can include the following:
\itemize{
\item{method}{Character. The method to mask clouds. Options: "cloud_prob" and "qa".}
\item{prob}{Numeric. Cloud probability threshold, between 0 and 100. Valid for 'cloud_prob' method. Default 60.}
\item{maskCirrus}{Logical. Whether to mask cirrus clouds. Valid for 'qa' method. Default TRUE.}
\item{maskShadows}{Logical. Whether to mask cloud shadows. Default TRUE.}
\item{scaledImage}{Logical. If TRUE, scale pixel values to <0,1>. Default FALSE.}
\item{dark}{Numeric. NIR threshold for potential cloud shadows, between 0-1. Default 0.15.}
\item{cloudDist}{Numeric. Max distance in meters to search for cloud shadows from cloud edges. Default 1000m.}
\item{buffer}{Numeric. Distance in meters to dilate cloud and shadow objects. Default 250m.}
\item{cdi}{Numeric. Cloud Displacement Index threshold, between <-1, 1>. Default NULL.}
}
For more information on parameters and methods, refer to relevant cloud masking literature and tutorials.
}
\examples{
\dontrun{
library(rgeeExtra)

ee_Initialize()

img <- ee$ImageCollection("COPERNICUS/S2_SR") \%>\%
  ee$ImageCollection$first() \%>\%
  ee$Image$Extra_maskClouds(prob = 75,buffer = 300,cdi = -0.5)

names(img)
}
}
