% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Image.R
\name{ee_Image_matchHistogram}
\alias{ee_Image_matchHistogram}
\title{Adjust the image's histogram to match a target image}
\usage{
ee_Image_matchHistogram(x, target, bands, geometry = NULL, maxBuckets = 256)
}
\arguments{
\item{x}{ee$Image to adjust.}

\item{target}{ee$Image image to match.}

\item{bands}{A dictionary of band names to match, with
source bands as keys and target bands as values.}

\item{geometry}{ee$Geometry. The region to match histograms in that
overlaps both images. If none is provided, the geometry of the source
image will be used.  Defaults is NULL.}

\item{maxBuckets}{Integer. The maximum number of buckets to use when building
histograms. Will be rounded to the nearest power of 2.  Default to 256.}
}
\value{
The adjusted image containing the matched source bands.
}
\description{
Adjust the image's histogram to match a target image
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

ee$ImageCollection$Dataset$NASA_GPM_L3_IMERG_V06$first() \%>\%
  ee_Image_getSTAC()
}
}
