% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_ImageColletion.R
\name{ee_ImageCollection_closest}
\alias{ee_ImageCollection_closest}
\title{Get the temporal nearest image}
\usage{
ee_ImageCollection_closest(x, date, tolerance = 1, unit = "month")
}
\arguments{
\item{x}{ee$ImageCollection. Image Collection from which to get
the closest image to the specified date.}

\item{date}{ee$Date or R date object. Date of interest. The function will
look for image closest to this date.}

\item{tolerance}{Numeric. Default 1. Filter the collection between
(date - tolerance: date + tolerance) before searching the closest
image. This speeds up the searching process for collections with a
high temporal resolution.}

\item{unit}{Character. Units for tolerance. Available units: "year",
"month", "week", "day", "hour", "minute" or "second". Default "month".}
}
\value{
An ee$Image closest to the specified date.
}
\description{
Gets the closest ee$Image to the specified date.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

roi <- ee$Geometry$Point(c(-79, -12))
ee$ImageCollection$Dataset$MODIS_006_MCD12Q1 \%>\%
  ee_ImageCollection_closest("2020-10-15") \%>\%
  ee$ImageCollection$first() \%>\%
  Map$addLayer()
}
}
