% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{extremeValues}
\alias{extremeValues}
\alias{ee_maxValue}
\alias{ee_minValue}
\title{Minimum and Maximum Values}
\usage{
`ee$Image$Extra_maxValue(x, ...)`

`ee$Image$Extra_minValue(x, ...)`
}
\arguments{
\item{x}{ee$Image to analyze.}

\item{...}{Additional arguments for specifying the mode and sampling.
See details for more information.}
}
\value{
A number representing the minimum or maximum value.
}
\description{
Returns the minimum or maximum value of an ee.Image. The return value will
be an approximation if the polygon (area of the ee.Image) contains
too many pixels at the native scale.
}
\details{
The \code{...} argument can include the following:
\itemize{
\item{mode}{Character. Indicates the geometry over which to reduce data. Options: "Rectangle" (default) or "Points".}
\item{sample_size}{Numeric. Number of points to be created. Relevant only if mode is "Points".}
}
The "Rectangle" mode uses the Image system:footprint, while the "Points" mode samples points over the Image system:footprint.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()
extra_Initialize()

image <- ee$ImageCollection$Dataset$LANDSAT_LC08_C01_T1$first()[["B1"]]
# max values
ee$Image$Extra_maxValue(image)
ee$Image$Extra_maxValue(image, mode = "Points", sample_size = 2)

# min values
ee$Image$Extra_minValue(image)
ee$Image$Extra_minValue(image, mode = "Points")
}
}
\concept{extremeValues}
