% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{extremeValues}
\alias{extremeValues}
\alias{ee_maxValue}
\alias{ee_minValue}
\title{Minimum and maximum values}
\usage{
ee_maxValue(image, mode = "Rectangle", sample_size = 1000)

ee_minValue(image, mode = "Rectangle", sample_size = 1000)
}
\arguments{
\item{image}{ee.Image}

\item{mode}{Character. Indicates the geometry over which
to reduce data (max). Two types are supported: "Rectangle" (by default)
which use the Image system:footprint, and "Points" which use points
sampled over the Image system:footprint.}

\item{sample_size}{Number of points to created. Ignore if mode is Rectangle.}
}
\value{
A number
}
\description{
Returns the minimum or maximum value of an ee.Image. The return value will
be an \strong{approximation} if the polygon (are of the ee.Image) contains
too many pixels at the native scale.
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

ee_Initialize()

image <- ee$Image$random()
# max values
maxValue(image)
maxValue(image, mode = "Points")

# min values
minValue(image)
minValue(image, mode = "Points")
}
}
\concept{extremeValues}
